pyproject.toml
[project]
name = "privicyagentbackend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sav.accharya", email = "sav.accharya@gmail.com" }
]
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "langchain-community>=0.0.20",
    "python-dotenv>=1.0.1",
    "langchain-google-genai>=0.0.9",
    "pydantic>=2.6.1",
    "langchain-google-community>=2.0.3",
    "google-generativeai>=0.8.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/privicyagentbackend"]

[tool.rye]
managed = true

[tool.rye.dev-dependencies]
autopep8 = ">=2.3.1"
